requirements.txt
Django==4.1.7
google_api_python_client==2.85.0
google_auth_oauthlib==1.0.0
protobuf==4.22.3
PyJWT==2.6.0
PyJWT==2.6.0
